Fjerner resterende bruk av abstrakt oppfølgingsklient. Test på at vi klarer å hente erBrukerUnderOppfølging da oppfølgingsklienten nå gjør henting, deserialisering og feilhåndtering inline
